DEVICE-523 Remove bad version of duplicate Temp/Humidity sensor
The 'SmartSense Temp/Humidity' device type contains issues that
don't exist in the 'SmartSense Temp/Humidity Sensor' type. Since
these types have the same finger print there is no reason to have
2 so I'm removing one.

The couple issues I noticed in the deleted version are:
* The battery reading is wrong
* The percentage indication is missing

I have migrated the 43 instances of this device type in production
and have commented out the fingerprint in prod so nobody else matches
it. See the comments in DEVICE-523 for more details on the cleanup.